requirements.txt
streamlit
requests
python-whois
tldextract
validators